separate some test AuthPlugins to single files
created AuthDeletePlugin and AuthCaseInsensitivePlugin under dokuwiki\test\mock
