+++ _test/tests/inc/auth_aclcheck_caseinsensitive.test.php
@@ -1,12 +1,6 @@
 <?php
 
-use dokuwiki\Extension\AuthPlugin;
-
-class auth_acl_caseinsensitive_auth extends AuthPlugin {
-    function isCaseSensitive() {
-        return false;
-    }
-}
+use dokuwiki\test\mock\AuthCaseInsensitivePlugin;
 
 class auth_acl_caseinsensitive_test extends DokuWikiTest {
     protected $oldAuth;
@@ -20,7 +14,7 @@ class auth_acl_caseinsensitive_test extends DokuWikiTest {
         $this->oldAuth    = $auth;
         $this->oldAuthAcl = $AUTH_ACL;
 
-        $auth = new auth_acl_caseinsensitive_auth();
+        $auth = new AuthCaseInsensitivePlugin();
     }
 
     function tearDown() {
